third_party/WebKit/LayoutTests/fast/viewport/viewport-48-expected.txt

ALERT: viewport size 3000x352 scale 1 with limits [0.25, 5] and userScalable true
